Refactor ObjC dependencies in definite_init_diagnostics tests
This change moves functionality that requires ObjC interop into a new file and
removes the XFAIL on Linux. Partially resolves SR-216.
